chore(triple-web-nextjs): 폴더 구조 정리
