--- packages/triple-web-nextjs/src/event-tracking-provider.tsx
@@ -1,32 +1,0 @@
-'use client'
-
-import {
-  EventTrackingProvider as EventTrackingProviderBase,
-  type EventTrackingProviderProps as EventTrackingProviderBaseProps,
-} from '@titicaca/triple-web'
-import { useSearchParams } from 'next/navigation'
-
-import { getEventTrackingUtm } from './providers/event-tracking-utm'
-
-export type EventTrackingProviderProps = Omit<
-  EventTrackingProviderBaseProps,
-  'utm'
->
-
-export function EventTrackingProvider({
-  children,
-  page,
-  onError,
-}: EventTrackingProviderProps) {
-  const searchParams = useSearchParams()
-
-  return (
-    <EventTrackingProviderBase
-      page={page}
-      utm={getEventTrackingUtm(searchParams)}
-      onError={onError}
-    >
-      {children}
-    </EventTrackingProviderBase>
-  )
-}
