+++ packages/triple-web-nextjs/src/index.ts
@@ -1,7 +1,2 @@
+export * from './initializers'
 export * from './providers'
-export * from './app-install-cta-modal-provider'
-export * from './build-triple-web-props'
-export * from './event-metadata-provider'
-export * from './event-tracking-provider'
-export * from './login-cta-modal-provider'
-export * from './triple-web'
