+++ packages/triple-web-nextjs/src/initializers/client-app.ts
