+++ packages/triple-web-nextjs/src/providers/event-metadata-provider.mdx
